64-bit integration: - upgrade of widgetlibrary SDK
git-svn-id: https://svn.code.sf.net/p/x-ivap/code/trunk@83 c9bdb542-074d-4938-be84-19bf99c82884

+++ WidgetLibrary/SampleCode/XPListBox.cpp
@@ -278,8 +278,8 @@ void XPListBoxDeleteItem(XPListBoxData_t *pListBoxData, int CurrentItem)
 int		XPListBoxProc(
 					XPWidgetMessage			inMessage,
 					XPWidgetID				inWidget,
-					long					inParam1,
-					long					inParam2)
+					intptr_t				inParam1,
+					intptr_t				inParam2)
 {
 //	static int SliderPosition, ScrollBarSlop, Min, Max, ScrollBarPageAmount, MaxListBoxItems;
 	static int ScrollBarSlop;
